+++ backend/services/albumphoto/mapper.go
@@ -7,7 +7,7 @@ import (
 )
 
 func toModel(data *viewmodels.AlbumPhotoDto, out *models.AlbumPhoto) {
-	out.ID = data.Id
+	out.ID = uint(data.Id)
 	out.Title = data.Title
 	out.Image = data.Image
 	out.Thumbnail = data.Thumbnail
@@ -17,7 +17,7 @@ func toModel(data *viewmodels.AlbumPhotoDto, out *models.AlbumPhoto) {
 }
 
 func toViewModel(in *models.AlbumPhoto, out *viewmodels.AlbumPhotoDto) {
-	out.Id = in.ID
+	out.Id = int(in.ID)
 	out.Title = in.Title
 	out.Image = in.Image
 	out.Thumbnail = in.Thumbnail
