+++ backend/services/article/mapper.go
@@ -22,9 +22,9 @@ func toModel(data *viewmodels.ArticleDto, out *models.Article) {
 	out.Body = data.Body
 	out.Image = data.Image
 	out.Thumbnail = data.Thumbnail
-	out.CreatedBy = data.UserId
-	out.UpdatedBy = data.UserId
-	out.ArticleTopicId = data.TopicId
+	out.CreatedBy = uint(data.UserId)
+	out.UpdatedBy = uint(data.UserId)
+	out.ArticleTopicId = uint(data.TopicId)
 
 	utils.FillCreated(data, out)
 	utils.FillUpdated(data, out)
@@ -36,9 +36,9 @@ func toViewModel(in *models.Article, out *viewmodels.ArticleDto) {
 	out.Body = in.Body
 	out.Image = in.Image
 	out.Thumbnail = in.Thumbnail
-	out.UserId = in.CreatedBy
+	out.UserId = int(in.CreatedBy)
 	out.CreatedAt = in.CreatedAt
-	out.TopicId = in.ArticleTopicId
+	out.TopicId = int(in.ArticleTopicId)
 
 	utils.FillCreated(in, out)
 	utils.FillUpdated(in, out)
