+++ backend/services/classroom/mapper.go
@@ -7,20 +7,20 @@ import (
 )
 
 func toModel(data *viewmodels.ClassroomDto, out *models.Classroom) {
-	out.ID = data.Id
+	out.ID = uint(data.Id)
 	out.Level = data.Level
 	out.Major = data.Major
-	out.Seq = data.Seq
+	out.Seq = uint8(data.Seq)
 
 	utils.FillCreated(data, out)
 	utils.FillUpdated(data, out)
 }
 
 func toViewModel(in *models.Classroom, out *viewmodels.ClassroomDto) {
-	out.Id = in.ID
+	out.Id = int(in.ID)
 	out.Level = in.Level
 	out.Major = in.Major
-	out.Seq = in.Seq
+	out.Seq = int(in.Seq)
 
 	utils.FillCreated(in, out)
 	utils.FillUpdated(in, out)
