+++ backend/viewmodels/album.go
@@ -3,21 +3,21 @@ package viewmodels
 import "backend/utils"
 
 type AlbumDto struct {
-	Id     uint   `query:"-" json:"id,omitempty" xml:"id,omitempty" form:"id,omitempty"`
-	UserId uint   `query:"userId,omitempty" json:"userId,omitempty" xml:"userId,omitempty" form:"userId,omitempty"`
+	Id     int    `query:"-" json:"id,omitempty" xml:"id,omitempty" form:"id,omitempty"`
+	UserId int    `query:"userId,omitempty" json:"userId,omitempty" xml:"userId,omitempty" form:"userId,omitempty"`
 	Title  string `query:"title,omitempty" json:"title,omitempty" xml:"title,omitempty" form:"title,omitempty"`
 	utils.Created
 	utils.Updated
 }
 
 func (p *AlbumDto) GetId() uint {
-	return p.Id
+	return uint(p.Id)
 }
 
 type AlbumPhotoDto struct {
-	AlbumId   uint   `query:"albumId" json:"albumId,omitempty" xml:"albumId,omitempty" form:"albumId,omitempty"`
+	AlbumId   int    `query:"albumId" json:"albumId,omitempty" xml:"albumId,omitempty" form:"albumId,omitempty"`
 	Title     string `query:"title" json:"title,omitempty" xml:"title,omitempty" form:"title,omitempty"`
-	Id        uint   `query:"-" json:"id,omitempty" xml:"id,omitempty" form:"id,omitempty"`
+	Id        int    `query:"-" json:"id,omitempty" xml:"id,omitempty" form:"id,omitempty"`
 	Image     string `query:"-" json:"original,omitempty" xml:"original,omitempty" form:"-"`
 	Thumbnail string `query:"-" json:"thumbnail,omitempty" xml:"thumbnail,omitempty" form:"-"`
 	utils.Created
@@ -25,7 +25,7 @@ type AlbumPhotoDto struct {
 }
 
 func (p *AlbumPhotoDto) GetId() uint {
-	return p.Id
+	return uint(p.Id)
 }
 
 type AlbumPhotoParam struct {
